// front-end/packages/shared/src/models/gender.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * OpenAPI specification - RING-Bookstore
 * OpenAPI documentation.
 * OpenAPI spec version: 2.0
 */

export type Gender = (typeof Gender)[keyof typeof Gender];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Gender = {
  MALE: "MALE",
  FEMALE: "FEMALE",
} as const;