// front-end/packages/shared/src/models/userRole.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * OpenAPI specification - RING-Bookstore
 * OpenAPI documentation.
 * OpenAPI spec version: 2.0
 */

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserRole = {
  ROLE_USER: "ROLE_USER",
  ROLE_SELLER: "ROLE_SELLER",
  ROLE_ADMIN: "ROLE_ADMIN",
  ROLE_GUEST: "ROLE_GUEST",
} as const;